+++ examples/scenes/src/TextLineLocaleScene.java
@@ -11,21 +11,16 @@ import io.github.humbleui.types.*;
 
 public class TextLineLocaleScene extends Scene {
     public final Paint textFill = new Paint().setColor(0xFF000000);
-    public final Font font;
-
-    public TextLineLocaleScene() {
-        // fontMgr.registerTypeface(Typeface.makeFromFile(file("fonts/NotoSans-Regular.ttf")));
-        // fontMgr.registerTypeface(Typeface.makeFromFile(file("fonts/NotoSansJP-Regular.otf")));
-        // fontMgr.registerTypeface(Typeface.makeFromFile(file("fonts/NotoSansSC-Regular.otf")));
-        // fontMgr.registerTypeface(Typeface.makeFromFile(file("fonts/NotoSansTC-Regular.otf")));
-        font = new Font().setSize(36f);
-    }
+    public final Font font = new Font().setSize(36f);
+    public final FontMgr fontMgr = FontMgr.getDefault();
 
     @Override
     public void draw(Canvas canvas, int width, int height, float dpi, int xpos, int ypos) {
+        canvas.translate(100, 100);
         try (var shaper = Shaper.makeShapeThenWrap();) {
             ShapingOptions opts = ShapingOptions.DEFAULT;
-            String text = "ja: 刃";
+            Typeface face = fontMgr.matchFamilyStyleCharacter(null, FontStyle.NORMAL, new String[] { "ja" }, "刃".codePointAt(0));
+            String text = "刃 ja " + face.getFamilyName();
             Iterator<LanguageRun> langIter = new TrivialLanguageRunIterator(text, "ja");
             try (ManagedString textUtf8 = new ManagedString(text);
                  FontMgrRunIterator fontIter = new FontMgrRunIterator(textUtf8, false, font, langIter, opts);
@@ -41,11 +36,13 @@ public class TextLineLocaleScene extends Scene {
                 
                 shaper.shape(textUtf8, fontIter, bidiIter, scriptIter, langIter, opts, width, runHandler);
                 try (TextLine line = runHandler.makeLine();) {
-                    canvas.drawTextLine(line, 20, 100, textFill);
+                    canvas.drawTextLine(line, 0, 0, textFill);
+                    canvas.translate(0, 50);
                 }
             }
 
-            text = "zh-Hans: 刃";
+            face = fontMgr.matchFamilyStyleCharacter(null, FontStyle.NORMAL, new String[] { "zh-Hans" }, "刃".codePointAt(0));
+            text = "刃 zh-Hans " + face.getFamilyName();
             Iterator<ScriptRun> scriptIter = new TrivialScriptRunIterator(text, "Hans");
             langIter = new TrivialLanguageRunIterator(text, "zh");
             try (ManagedString textUtf8 = new ManagedString(text);
@@ -56,11 +53,13 @@ public class TextLineLocaleScene extends Scene {
                 
                 shaper.shape(textUtf8, fontIter, bidiIter, scriptIter, langIter, opts, width, runHandler);
                 try (TextLine line = runHandler.makeLine();) {
-                    canvas.drawTextLine(line, 20, 200, textFill);
+                    canvas.drawTextLine(line, 0, 0, textFill);
+                    canvas.translate(0, 50);
                 }
             }
 
-            text = "zh-Hant: 刃";
+            face = fontMgr.matchFamilyStyleCharacter(null, FontStyle.NORMAL, new String[] { "zh-Hant" }, "刃".codePointAt(0));
+            text = "刃 zh-Hant " + face.getFamilyName();
             langIter = new TrivialLanguageRunIterator(text, "zh");
             scriptIter = new TrivialScriptRunIterator(text, "Hant");
             try (ManagedString textUtf8 = new ManagedString(text);
@@ -70,12 +69,14 @@ public class TextLineLocaleScene extends Scene {
             {
                 shaper.shape(textUtf8, fontIter, bidiIter, scriptIter, langIter, opts, width, runHandler);
                 try (TextLine line = runHandler.makeLine();) {
-                    canvas.drawTextLine(line, 20, 300, textFill);
+                    canvas.drawTextLine(line, 0, 0, textFill);
+                    canvas.translate(0, 50);
                 }
             }
 
             try (var line = shaper.shapeLine("ja: 刃, zh-Hans: 刃, zh-Hant: 刃", font)) {
-                canvas.drawTextLine(line, 20, 400, textFill);
+                canvas.drawTextLine(line, 0, 0, textFill);
+                canvas.translate(0, 50);
             }
         }
     }