+++ platform/cc/paragraph/ParagraphBuilder.cc
@@ -50,13 +50,6 @@ extern "C" JNIEXPORT void JNICALL Java_io_github_humbleui_skija_paragraph_Paragr
     instance->addPlaceholder({width, height, static_cast<PlaceholderAlignment>(alignment), static_cast<TextBaseline>(baselinePosition), baseline});
 }
 
-extern "C" JNIEXPORT void JNICALL Java_io_github_humbleui_skija_paragraph_ParagraphBuilder__1nSetParagraphStyle
-  (JNIEnv* env, jclass jclass, jlong ptr, jlong stylePtr) {
-    ParagraphBuilder* instance = reinterpret_cast<ParagraphBuilder*>(static_cast<uintptr_t>(ptr));
-    ParagraphStyle* style = reinterpret_cast<ParagraphStyle*>(static_cast<uintptr_t>(stylePtr));
-    instance->setParagraphStyle(*style);
-}
-
 extern "C" JNIEXPORT jlong JNICALL Java_io_github_humbleui_skija_paragraph_ParagraphBuilder__1nBuild
   (JNIEnv* env, jclass jclass, jlong ptr) {
     ParagraphBuilder* instance = reinterpret_cast<ParagraphBuilder*>(static_cast<uintptr_t>(ptr));